api/requirements.txt
psycopg2-binary==2.9.10 